Move getDests to gameCtrl as setDests
